Wrote LPA*, which is the discrete version of the backwards dstar lite st class, and got it to perform similarly to normal cbsmp

+++ 16891_project_ws/src/mamp_planning/include/mamp_planning/edge.hpp
@@ -18,7 +18,7 @@ public:
     void setTraversalTime(double t);
     double getTraversalTime();
     void setDivisions(double divisions);
-    double getDivisions();
+    unsigned int getDivisions();
     std::shared_ptr<std::vector<std::vector<double>>> getVertexPositions();
     std::shared_ptr<std::vector<std::vector<double>>> getVertexPositionsInGivenOrder(std::shared_ptr<Vertex> v1, std::shared_ptr<Vertex> v2);
     std::pair<double, std::vector<double>> getMagnitudeAndUnitVector(std::shared_ptr<Vertex> start_vertex);
@@ -28,7 +28,7 @@ private:
     double cost_;
     bool valid_;
     double traversal_time_;
-    double divisions_;
+    unsigned int divisions_;
     
     std::unordered_map<unsigned int, std::shared_ptr<Vertex>> vertices_;
 };
